don't convert to np.float - can double data size!  (very bad for large cubes)

+++ spectral_cube/masks.py
@@ -119,7 +119,7 @@ class MaskBase(object):
         This is an internal method used by :class:`SpectralCube`.
         Users should use the property :meth:`MaskBase.filled_data`
         """
-        sliced_data = data[view].copy().astype(np.float)
+        sliced_data = data[view].copy().astype(data.dtype)
         ex = self.exclude(data=data, wcs=wcs, view=view)
         sliced_data[ex] = fill
         return sliced_data
